requirements.txt
django 
djangorestframework
environs